(#273) Move quote commands to src/cmd/quote.py
